clean up the code and move login logic to login.js
